+++ Runtime/ResourceManager/ResourceLocations/ResourceLocationBase.cs
@@ -109,4 +109,33 @@ namespace UnityEngine.ResourceManagement.ResourceLocations
                 m_DependencyHashCode = m_DependencyHashCode * 31 + d.Hash(typeof(object));
         }
     }
+
+    internal class LocationWrapper : IResourceLocation
+    {
+        IResourceLocation m_InternalLocation;
+        public LocationWrapper(IResourceLocation location)
+        {
+            m_InternalLocation = location;
+        }
+        public string InternalId => m_InternalLocation.InternalId;
+
+        public string ProviderId => m_InternalLocation.ProviderId;
+
+        public IList<IResourceLocation> Dependencies => m_InternalLocation.Dependencies;
+
+        public int DependencyHashCode => m_InternalLocation.DependencyHashCode;
+
+        public bool HasDependencies => m_InternalLocation.HasDependencies;
+
+        public object Data => m_InternalLocation.Data;
+
+        public string PrimaryKey => m_InternalLocation.PrimaryKey;
+
+        public Type ResourceType => m_InternalLocation.ResourceType;
+
+        public int Hash(Type resultType)
+        {
+            return m_InternalLocation.Hash(resultType);
+        }
+    }
 }